+++ src/main/java/nextstep/auth/interceptor/TokenAuthenticationInterceptor.java
@@ -1,7 +1,10 @@
-package nextstep.auth.token;
+package nextstep.auth.interceptor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.authentication.AuthenticationException;
+import nextstep.auth.exception.AuthenticationException;
+import nextstep.auth.token.JwtTokenProvider;
+import nextstep.auth.token.TokenRequest;
+import nextstep.auth.token.TokenResponse;
 import nextstep.member.application.LoginMemberService;
 import nextstep.member.domain.LoginMember;
 import org.springframework.http.MediaType;
